Added Mui.css 0.4.6 as vendored asset
Tried multiple configuration: Bower, NPM. Could not make work correctly any of them. In the best case, CSS was ok, JS was compiled (when trying with NPM), but I could not make it fully work. As if NPM build for Mui.css was not as complete as the pre-built mui.js file.