[Bugfix] TeamMemberList update
Fixed the way the TeamMemberList is updated on ReceivedIssues action to ensure we add new team member if presents in the issues and we correctly update the assignments over all team members in the list. Previously, was not adding the new team members correctly and team members not present in the received issues were not getting their assignments updated.